tests for Store, testing coverage works

+++ src/main/java/sakila_api/sakila/Film.java
@@ -85,7 +85,7 @@ public class Film {
 
     public int getLanguage_id(){return this.language_id;}
 
-    public double getRental_duration(){return this.rental_duration;}
+    public int getRental_duration(){return this.rental_duration;}
 
     public double getRental_rate(){return this.rental_rate;}
